+++ src/app/admin/page.tsx
@@ -1,8 +1,8 @@
 
 export default function AdminPage() {
     return (
-        <div>
-            admin page
+        <div className="text-2xl text-white text-bold">
+            sup ai club pepes hows it going?
         </div>
     )
 }
