+++ src/app/layout.tsx
@@ -5,6 +5,7 @@ import NavBar from "~/components/nav/NavBar";
 import { Providers } from "~/app/providers";
 import { createClient } from "~/utils/supabase/server";
 import { getProfileType } from "~/server/db/queries/profiles";
+import { Toaster } from "react-hot-toast";
 
 export const metadata = {
   title: "MSU AI Club",
@@ -26,7 +27,8 @@ export default async function RootLayout({
   return (
     <html lang="en" className={`${GeistSans.variable}`}>
       <body className="bg-slate-700 h-screen w-screen">
-        <Providers>
+        <Providers>  
+          <Toaster />
           <NavBar userType={userType} />
           {children}
         </Providers>
