+++ src/app/posts/[postId]/edit/editForm.tsx
@@ -1,22 +1,25 @@
 "use client";
 
 import { useRouter } from "next/navigation";
-import { createPost } from "~/server/actions/post";
-import PostForm from "../postForm";
+import { updatePost } from "~/server/actions/post";
+import PostForm from "../../postForm";
+import { Post } from "~/types/posts";
 
 
-export default function NewPostForm({
-    supaId
+export default function EditPostForm({
+    supaId,
+    post
 } : {
-    supaId: string | undefined
+    supaId: string | undefined,
+    post: Post
 }) {
     const router = useRouter();
 
     const handleSubmit = async (name: string, content: string) => {
-        const res = await createPost(supaId!, name, content);
+        const res = await updatePost(supaId!, post?.id, name, content);
 
         if (res !== null) {
-            router.push(`/member/posts/${res}`);
+            router.push(`/posts/${res}`);
         } else {
             alert("Could not create post");
         }
@@ -26,6 +29,8 @@ export default function NewPostForm({
         <div>
             <PostForm 
             supaId={supaId}
+            prevName={post?.name}
+            prevContent={post?.content}
             handleSubmit={handleSubmit}
             />
         </div>
