+++ src/app/posts/[postId]/page.tsx
@@ -1,4 +1,8 @@
+import { Button } from "@nextui-org/react";
 import { getPostById } from "~/server/db/queries/posts";
+import DeleteButton from "./deletPostButton";
+import { createClient } from "~/utils/supabase/server";
+import Link from "next/link";
 
 export default async function PostPage({
     params
@@ -7,17 +11,29 @@ export default async function PostPage({
         postId: string
     }
 }) {
+    const supabase = createClient();
+
+    const { data } = await supabase.auth.getUser();
+
     const post = await getPostById(params.postId);
 
-    if (post === null) {
+    if (!post) {
         return <div>Post not found</div>;
     }
 
     return (
+        <>
         <div>
-            Post
-            {post.name}
+            Admin Post 
+            {post.id} 
+            {post.name} 
             {post.content}
         </div>
+        <DeleteButton 
+        postId={post.id} 
+        userId={data?.user?.id}
+        />
+        <Link href={`/posts/${post.id}/edit`}>Edit</Link>
+        </>
     )
 }
