+++ src/app/posts/makePostButton.tsx
@@ -5,7 +5,7 @@ import Link from "next/link";
 
 export default function MakePostButton() {
     return (
-        <Button as={Link} href="/member/posts/new">
+        <Button as={Link} href="/posts/new">
             Make post
         </Button>
     );
