+++ src/app/posts/page.tsx
@@ -1,28 +1,29 @@
 import Link from "next/link";
-import { getPosts } from "~/server/db/queries/posts";
+import { getUserPosts } from "~/server/db/queries/posts";
 import type { Post } from "~/types/posts";
-import { getProfileType } from "~/server/db/queries/profiles";
 import { createClient } from "~/utils/supabase/server";
+import MakePostButton from "./makePostButton";
+import { getProfileType } from "~/server/db/queries/profiles";
 
 export default async function PostsPage() {
-    const posts = await getPosts();
-    
     const supabase = createClient();
 
-    const { data } = await supabase.auth.getUser();  
+    const { data } = await supabase.auth.getUser();
 
     const userType: string | null = await getProfileType(data.user?.id);
 
-    console.log(userType);
+    const posts = await getUserPosts(data?.user?.id ?? null);
     
+
     return (
         <div>
             Posts
             {posts.map((post: Post) => (
-                <Link href={`/member/posts/${post.id}`} key={post.id}>
-                    {post.name}
-                </Link>
+                <div key={post.id}>
+                    <Link href={`/posts/${post.id}`}>{post.name}</Link>
+                </div>
             ))}
+            {  userType === "admin" && <MakePostButton /> }
         </div>
     )
 }
